+++ common/src/main/kotlin/com/gitlab/kordlib/common/entity/Permission.kt
@@ -1,7 +1,6 @@
 package com.gitlab.kordlib.common.entity
 
 import kotlinx.serialization.*
-import kotlinx.serialization.internal.IntDescriptor
 
 @Serializable(with = Permissions.Companion::class)
 class Permissions constructor(val code: Int) {
@@ -91,6 +90,7 @@ enum class Permission(val code: Int = 0) {
     ReadMessageHistory(0x00010000),
     MentionEveryone(0x00020000),
     UseExternalEmojis(0x00040000),
+    ViewGuildInsights(0x00080000),
     Connect(0x00100000),
     Speak(0x00200000),
     MuteMembers(0x00400000),
@@ -103,5 +103,5 @@ enum class Permission(val code: Int = 0) {
     ManageRoles(0x10000000),
     ManageWebhooks(0x20000000),
     ManageEmojis(0x40000000),
-    All(0x7ED7FDFF)
+    All(0x7FFFFDFF)
 }