+++ gateway/src/main/kotlin/com/gitlab/kordlib/gateway/DefaultGateway.kt
@@ -16,8 +16,7 @@ import io.ktor.util.error
 import kotlinx.atomicfu.AtomicRef
 import kotlinx.atomicfu.atomic
 import kotlinx.atomicfu.update
-import kotlinx.coroutines.CancellationException
-import kotlinx.coroutines.ObsoleteCoroutinesApi
+import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.BroadcastChannel
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.channels.ReceiveChannel
@@ -25,7 +24,6 @@ import kotlinx.coroutines.channels.sendBlocking
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
-import kotlinx.coroutines.withTimeoutOrNull
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonConfiguration
 import mu.KotlinLogging
@@ -99,7 +97,8 @@ class DefaultGateway(private val data: DefaultGatewayData) : Gateway {
         InvalidSessionHandler(events) { restart(it) }
     }
 
-    override suspend fun start(configuration: GatewayConfiguration) {
+    //running on default dispatchers because ktor does *not* like running on an EmptyCoroutineContext from main
+    override suspend fun start(configuration: GatewayConfiguration): Unit = withContext(Dispatchers.Default) {
         resetState(configuration)
 
         while (data.reconnectRetry.hasNext && state.value is State.Running) {
@@ -266,10 +265,15 @@ class DefaultGateway(private val data: DefaultGatewayData) : Gateway {
     }
 
     @Suppress("EXPERIMENTAL_API_USAGE")
-    private suspend fun sendUnsafe(command: Command) = stateMutex.withLock {
+    private suspend fun sendUnsafe(command: Command) {
         data.sendRateLimiter.consume()
         val json = Json.stringify(Command.Companion, command)
-        if (command is Identify) defaultGatewayLogger.trace { "Gateway >>> Identify" }
+        if (command is Identify) {
+            defaultGatewayLogger.trace {
+                val copy = command.copy(token = "token")
+                "Gateway >>> ${Json.stringify(Command.Companion, copy)}"
+            }
+        }
         else defaultGatewayLogger.trace { "Gateway >>> $json" }
         socket.send(Frame.Text(json))
     }
@@ -286,7 +290,7 @@ class DefaultGateway(private val data: DefaultGatewayData) : Gateway {
     }
 }
 
-internal val GatewayConfiguration.identify get() = Identify(token, IdentifyProperties(os, name, name), false, 50, shard, presence)
+internal val GatewayConfiguration.identify get() = Identify(token, IdentifyProperties(os, name, name), false, 50, shard, presence, intents)
 
 internal val os: String get() = System.getProperty("os.name")