+++ rest/src/main/kotlin/com/gitlab/kordlib/rest/service/GuildService.kt
@@ -198,17 +198,6 @@ class GuildService(requestHandler: RequestHandler) : RestService(requestHandler)
         body(GuildIntegrationCreateRequest.serializer(), integration)
     }
 
-    @Deprecated("use the inline builder instead", ReplaceWith("""modifyGuildIntegration(guildId, integrationId) {
-        | this.expireBehavior = integration.expireBehavior,
-        | this.expirePeriod = integration.expirePeriod,
-        | this.emoticons = integration.emoticons
-        |})"""), DeprecationLevel.WARNING)
-    suspend fun modifyGuildIntegration(guildId: String, integrationId: String, integration: GuildIntegrationModifyRequest) = call(Route.GuildIntegrationPatch) {
-        keys[Route.GuildId] = guildId
-        keys[Route.IntegrationId] = integrationId
-        body(GuildIntegrationModifyRequest.serializer(), integration)
-    }
-
     suspend inline fun modifyGuildIntegration(guildId: String, integrationId: String, builder: IntegrationModifyBuilder.() -> Unit) = call(Route.GuildIntegrationPatch) {
         keys[Route.GuildId] = guildId
         keys[Route.IntegrationId] = integrationId